+++ app/Http/Controllers/TrajetController.php
@@ -11,6 +11,7 @@ use App\Models\Trajet;
 use App\Models\Carburant;
 use App\Models\Chauffeur;
 use App\Models\Itineraire;
+use App\Models\Reservation;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 use Illuminate\Http\JsonResponse;
@@ -37,22 +38,27 @@ class TrajetController extends Controller
     {
         // Validation des données reçues
 
+        
         $data = $request->validate([
             "camion_id" => ['required', 'numeric', 'exists:camions,id'],
             "etat" => ['required', Rule::in(Trajet::getEtat())],
             "chauffeur" => ['nullable', 'exists:chauffeurs,id'],
             "date_heure_depart" => ['required', 'date'],
             "date_heure_arrivee" => ['nullable', 'date'],
-            "carburantRestant" => ['nullable', 'numeric', 'min:1', 'max:500'],
+            "carburantRestant" => ['nullable', 'numeric'],
         ]);
 
-        $date_depart = Carbon::parse($request->date_heure_depart, 'EAT');
-        $date_arrivee = $request->date_heure_arrivee === null ? null : Carbon::parse($request->date_heure_arrivee, 'EAT');
 
+
+        $date_depart = Carbon::parse(date("Y-m-d H:i:s", strtotime($request->date_heure_depart)), 'EAT');
+        $date_arrivee = $request->date_heure_arrivee === null ? null : Carbon::parse(date("Y-m-d H:i:s", strtotime($request->date_heure_arrivee)), 'EAT');
+
+        $carburant = collect();
         // Verifier si le camion a un trajet en cours ou non
         $camion = Camion::findOrFail($request->camion_id);
-        $chauffeur = Chauffeur::findOrFail($request->chauffeur);
+        $chauffeur = Chauffeur::find($request->chauffeur);
 
+        
         if($camion->estDispoEntre($date_depart, $date_arrivee) !== true){
 
             $request->session()->flash("notification", [
@@ -66,23 +72,26 @@ class TrajetController extends Controller
         if ($request->chauffeur !== null)
         {
             $chauffeur = Chauffeur::findOrFail($request->chauffeur);
-
+           
+           /* $msg = "Aucun chauffeur n'a été selectionnée";
             if (!$chauffeur->estDispoEntre($date_depart, $date_arrivee))
             {
-                $request->session()->flash("notification", [
-                    "value" => "Chauffeur non disponible entre les dates que vous avez selectionné" ,
-                    "status" => "error"
-                ]);
-
-                return redirect()->route('camion.voir', ['camion' => $camion->id, 'tab' => 2]);
+                $msg = "Chauffeur non disponible entre les dates que vous avez selectionné";
             }
+
+            $request->session()->flash("notification", [
+                "value" =>  $msg ,
+                "status" => "error"
+            ]);
+
+            return redirect()->route('camion.voir', ['camion' => $camion->id, 'tab' => 2]);*/
         }
         else
         {
             if ($request->etat !== Trajet::getEtat(0))
             {
                 $request->session()->flash("notification", [
-                    "value" => "Vous devez selectionner au moins un chauffeur pour un trajet non a prévoir" ,
+                    "value" => "Vous devez selectionner au moins un chauffeur pour un trajet a prévoir" ,
                     "status" => "error"
                 ]);
 
@@ -95,7 +104,7 @@ class TrajetController extends Controller
         {
             $request->session()->flash("notification", [
                 "value" => "Le camion a encore un trajet en cours" ,
-                "status" => "success"
+                "status" => "error"
             ]);
 
             return redirect()->route('camion.voir', ['camion' => $camion->id, 'tab' => 2]);
@@ -104,7 +113,7 @@ class TrajetController extends Controller
         if (Carbon::now('EAT')->greaterThanOrEqualTo($date_depart) AND $request->etat === Trajet::getEtat(0))
         {
             $request->session()->flash("notification", [
-                "value" => "La date de depart doit être spérieur a ce moment précis si la status est aprévoir" ,
+                "value" => "La date de depart doit être supérieur a ce moment précis si le statut est à prévoir" ,
                 "status" => "error"
             ]);
 
@@ -156,15 +165,17 @@ class TrajetController extends Controller
         if ($date_arrivee !== null AND $date_depart->greaterThan($date_arrivee))
         {
             $request->session()->flash("notification", [
-                "value" => "La date de depart doit etrer inférieur a la date d\'arrivée" ,
+                "value" => "La date de depart doit être inférieur a la date d\'arrivée" ,
                 "status" => "error"
             ]);
 
             return redirect()->route('camion.voir', ['camion' => $camion->id, 'tab' => 2]);
         }
 
-        // Verifier si la status est terminé et que la carbburant restant n'est pas nulle
-        if ($request->etat === Trajet::getEtat(2) AND $request->carburantRestant === null)
+        // Verifier si la status est terminé et que la carburant restant n'est pas nulle
+        $carburant_total = $request->etat === Trajet::getEtat(2) ? doubleval($camion->stockCarburant()) -  doubleval($request->carburantRestant) : null;
+        $carburant_depart = ( $camion->stockCarburant() >= doubleval($request->carburantRestant) && doubleval($request->carburantRestant) > 0 )  ? doubleval($request->carburantRestant) : doubleval($camion->stockCarburant());
+        if (($request->etat === Trajet::getEtat(1) || $request->etat === Trajet::getEtat(2)) AND $request->carburantRestant === null)
         {
             $request->session()->flash("notification", [
                 "value" => "Veillez remplir la quantité de carburant restant" ,
@@ -173,20 +184,50 @@ class TrajetController extends Controller
 
             return redirect()->route('camion.voir', ['camion' => $camion->id, 'tab' => 2]);
             
-        }else if($request->etat === Trajet::getEtat(2)){
+        }
+        
+        else if($request->etat === Trajet::getEtat(1) && $carburant_depart == 0){
+
+
+            $request->session()->flash("notification", [
+                "value" => "Le carburant du véhicule est encore insuffisant" ,
+                "status" => "error"
+            ]);
+            
+            return redirect()->route('camion.voir', ['camion' => $camion->id, 'tab' => 2]);
+        }
+        
+        else if($request->etat === Trajet::getEtat(2)){
+
+            if($carburant_total < 0 ){
+                
+                $request->session()->flash("notification", [
+                    "value" => "La quantité de carburant que vous avez saisi est superieur au stock actuel" ,
+                    "status" => "error"
+                ]);
+                
+                return redirect()->route('camion.voir', ['camion' => $camion->id, 'tab' => 2]);
+            }
 
             $CarburantSortie = doubleval($camion->CarburantRestant()) - doubleval($request->carburantRestant);
 
-            Carburant::create([
-                "quantite" => $CarburantSortie,
-                "flux" => 1,
-                "date" => $date_arrivee,
-                "camion_id" => $camion->id
-            ]);
+            Carburant::where("id", $trajet->carburant_id)->delete();
+
+            if($CarburantSortie > 0){
+
+                $carburant = Carburant::create([
+                    "quantite" => $CarburantSortie,
+                    "flux" => 1,
+                    "date" => $date_arrivee,
+                    "camion_id" => $camion->id
+                ]);
+            }
+
         }
 
-        $depart = ucfirst($itineraires[0]['nom_itineraire']);
-        $arrivee = ucfirst(end($itineraires)['nom_itineraire']);
+        $depart = ucfirst($itineraires[0]['nom']);
+        $arrivee = ucfirst(end($itineraires)['nom']);
+
 
         $trajet = new Trajet([
             'depart' => $depart,
@@ -196,6 +237,10 @@ class TrajetController extends Controller
             'etat' => $request->etat,
             'camion_id' => intval($request->camion_id),
             'chauffeur_id' => $request->chauffeur === null ? null : intval($request->chauffeur),
+            'carburant_depart' => ( $request->etat == Trajet::getEtat(1) || $request->etat == Trajet::getEtat(2) ) ? $carburant_depart : null ,
+            'carburant_total' => $carburant_total,
+            'carburant_id' => isset($carburant->id) === true ? $carburant->id : null
+
         ]);
 
         if ($trajet->save())
@@ -208,7 +253,7 @@ class TrajetController extends Controller
             foreach ($itineraires as $itineraire)
             {
                 $itineraire = Itineraire::create([
-                    'nom' => $itineraire['nom_itineraire'],
+                    'nom' => $itineraire['nom'],
                     'id_trajet' => $trajet->id,
                 ]);
             }
@@ -238,6 +283,7 @@ class TrajetController extends Controller
     */
     public function modifier(Trajet $trajet) : JsonResponse
     {
+        
         $itineraires = $trajet->itineraires;
 
         return response()->json([
@@ -265,12 +311,16 @@ class TrajetController extends Controller
             "chauffeur" => ['nullable', 'exists:chauffeurs,id'],
             "date_heure_depart" => ['required', 'date'],
             "date_heure_arrivee" => ['nullable', 'date'],
-            "itineraire" => ["required", "sometimes"]
+            "itineraire" => ["required", "sometimes"],
+            "carburantRestant" => ['nullable', 'numeric']
         ]);
 
-        $date_depart = Carbon::parse($request->date_heure_depart, 'EAT');
-        $date_arrivee = $request->date_heure_arrivee === null ? null : Carbon::parse($request->date_heure_arrivee, 'EAT');
+        $date_depart = Carbon::parse(date("Y-m-d H:i:s", strtotime($request->date_heure_depart)), 'EAT');
+        $date_arrivee = $request->date_heure_arrivee === null ? null : Carbon::parse(date("Y-m-d H:i:s", strtotime($request->date_heure_arrivee)), 'EAT');
+
         $camion = Camion::findOrFail($request->camion_id);
+
+        $carburant = collect();
 
         if($camion->estDispoEntre($date_depart, $date_arrivee, $trajet ) !== true){
 
@@ -392,23 +442,126 @@ class TrajetController extends Controller
             }
         }
 
+        
         $itineraires = json_decode($data['itineraire'], true);
+        
+        // Si l'itinéraire est inférieur a deux
+        if (count($itineraires) < 2)
+        {
+            $request->session()->flash("notification", [
+                "value" => "Vous devez choisir au moins deux itinéraires" ,
+                "status" => "error"
+            ]);
 
+            return redirect()->route('camion.voir', ['camion' => $camion->id, 'tab' => 2]);
+        }
+        
         $depart = $itineraires[0]['nom'];
         $arrivee = end($itineraires)['nom'];
         $etat = $request->etat;
 
-        $update = $trajet->update([
+        // Verifier si la status est terminé et que la carburant restant n'est pas nulle
+        $carburant_total = $request->etat === Trajet::getEtat(2) ? doubleval($camion->stockCarburant()) -  doubleval($request->carburantRestant) : null;
+        $carburant_depart = $camion->stockCarburant() >= doubleval($request->carburantRestant) ? doubleval($request->carburantRestant) : 0;
+
+        if (($request->etat === Trajet::getEtat(1) || $request->etat === Trajet::getEtat(2) ) AND $request->carburantRestant === null)
+        {
+            $request->session()->flash("notification", [
+                "value" => "Veillez remplir la quantité de carburant restant" ,
+                "status" => "error"
+            ]);
+
+            return redirect()->route('camion.voir', ['camion' => $camion->id, 'tab' => 2]);
+            
+        }
+        
+        else if($request->etat === Trajet::getEtat(1) && $carburant_depart == 0){
+
+            
+            $request->session()->flash("notification", [
+                "value" => "Le carburant du véhicule est encore insuffisant" ,
+                "status" => "error"
+            ]);
+            
+            return redirect()->route('camion.voir', ['camion' => $camion->id, 'tab' => 2]);
+        }
+        
+        else if($request->etat === Trajet::getEtat(2)){
+
+            if($carburant_total < 0 ){
+                
+                $request->session()->flash("notification", [
+                    "value" => "La quantité de carburant que vous avez saisi est superieur au stock" ,
+                    "status" => "error"
+                ]);
+                
+                return redirect()->route('camion.voir', ['camion' => $camion->id, 'tab' => 2]);
+            }
+
+            $CarburantSortie = doubleval($camion->CarburantRestant()) - doubleval($request->carburantRestant);
+            
+            Carburant::where("id", $trajet->carburant_id)->delete();
+
+            if($CarburantSortie > 0){
+
+                $carburant = Carburant::create([
+                    "quantite" => $CarburantSortie,
+                    "flux" => 1,
+                    "date" => $date_arrivee,
+                    "camion_id" => $camion->id
+                ]);
+            }
+        }
+
+ 
+        $_trajet = [
             'depart' => $depart,
             'date_heure_depart' => $date_depart->toDateTimeString(),
             'arrivee' => $arrivee,
             'date_heure_arrivee' => $date_arrivee,
             'etat' => $etat,
             'chauffeur_id' => $request->chauffeur === null ? null : intval($request->chauffeur),
-        ]);
+            'carburant_id' => isset($carburant->id) === true ? $carburant->id : null
+
+        ];
+
+        if($request->etat === Trajet::getEtat(1)){
+            $_trajet["carburant_depart"] = $carburant_depart;
+        }
+
+        if($request->etat === Trajet::getEtat(2)){
+            $_trajet["carburant_total"] = $carburant_total;
+        }
+        
+        $update = $trajet->update($_trajet);
 
         if ($update)
         {
+            if( isset($trajet->reservation->id) === true ){
+
+                if($trajet->etat == Trajet::getEtat(2)){
+                    $trajet->reservation->status = Reservation::STATUS[2];
+                    $trajet->reservation->update();
+                }
+    
+                if($trajet->etat == Trajet::getEtat(3)){
+                    $trajet->reservation->status = Reservation::STATUS[6];
+                    $trajet->reservation->update();
+                }
+            }
+
+
+            //Supprimer les itineraire existant
+            $trajet->viderTrajet();
+            
+            foreach ($itineraires as $itineraire)
+            {
+                $itineraire = Itineraire::create([
+                    'nom' => $itineraire['nom'],
+                    'id_trajet' => $trajet->id,
+                ]);
+            }
+
             $request->session()->flash("notification", [
                 "value" => "Mise a jour avec success." ,
                 "status" => "success"
@@ -435,24 +588,35 @@ class TrajetController extends Controller
     */
     public function supprimer(Request $request, Trajet $trajet) : RedirectResponse
     {
-        $trajet->itineraires()->delete();
-        $camion_id = $trajet->camion_id;
-        $delete = $trajet->delete();
 
-        if ($delete)
-        {
+        $camion_id = $trajet->camion_id;
+
+        if(isset($trajet->reservation->id) === false){
+            $trajet->itineraires()->delete();
+            $delete = $trajet->delete();
+    
+            if ($delete)
+            {
+                $request->session()->flash("notification", [
+                    "value" => "Trajet supprimé" ,
+                    "status" => "success"
+                ]);
+            }
+            else
+            {
+                $request->session()->flash("notification", [
+                    "value" => "Impossible de supprimer le trajet" ,
+                    "status" => "error"
+                ]);
+            }
+        }else{
+
             $request->session()->flash("notification", [
-                "value" => "Trajet supprimé" ,
-                "status" => "success"
-            ]);
-        }
-        else
-        {
-            $request->session()->flash("notification", [
-                "value" => "Impossible de supprimer le trajet" ,
+                "value" => "Ce trajet n'est pas suppressible" ,
                 "status" => "error"
             ]);
         }
+
 
         return redirect()->route('camion.voir', ['camion' => $camion_id, 'tab' => 2]);
